+++ bsd/sys/resourcevar.h
@@ -78,11 +78,11 @@ struct pstats {
 	struct uprof {			/* profile arguments */
 		struct uprof *pr_next;  /* multiple prof buffers allowed */
 		caddr_t	pr_base;	/* buffer base */
-		u_long	pr_size;	/* buffer size */
-		u_long	pr_off;		/* pc offset */
-		u_long	pr_scale;	/* pc scaling */
-		u_long	pr_addr;	/* temp storage for addr until AST */
-		u_long	pr_ticks;	/* temp storage for ticks until AST */
+		u_int32_t	pr_size;	/* buffer size */
+		u_int32_t	pr_off;		/* pc offset */
+		u_int32_t	pr_scale;	/* pc scaling */
+		u_int32_t	pr_addr;	/* temp storage for addr until AST */
+		u_int32_t	pr_ticks;	/* temp storage for ticks until AST */
 	} p_prof;
 #define	pstat_endzero	p_start
 
@@ -127,7 +127,7 @@ struct plimit {
                 (proc_is64bit((p)) ? (p)->p_stats->user_p_prof.pr_ticks \
                                    : (p)->p_stats->p_prof.pr_ticks))
 
-void	 addupc_intr(struct proc *p, u_long pc, u_int ticks);
+void	 addupc_intr(struct proc *p, uint32_t pc, u_int ticks);
 void	 addupc_task(struct proc *p, user_addr_t pc, u_int ticks);
 void	 calcru(struct proc *p, struct timeval *up, struct timeval *sp,
 	    struct timeval *ip);
