+++ iokit/Kernel/IOFilterInterruptEventSource.cpp
@@ -42,19 +42,19 @@ HISTORY
 #define IOTimeTypeStampS(t)						\
 do {									\
     IOTimeStampStart(IODBG_INTES(t),					\
-                     (unsigned int) this, (unsigned int) owner);	\
+                     (uintptr_t) this, (uintptr_t) owner);	\
 } while(0)
 
 #define IOTimeTypeStampE(t)						\
 do {									\
     IOTimeStampEnd(IODBG_INTES(t),					\
-                   (unsigned int) this, (unsigned int) owner);		\
+                   (uintptr_t) this, (uintptr_t) owner);		\
 } while(0)
 
 #define IOTimeStampLatency()						\
 do {									\
     IOTimeStampEnd(IODBG_INTES(IOINTES_LAT),				\
-                   (unsigned int) this, (unsigned int) owner);		\
+                   (uintptr_t) this, (uintptr_t) owner);		\
 } while(0)
 
 #else /* !KDEBUG */
@@ -161,7 +161,7 @@ IOTimeTypeStampS(IOINTES_INTCTXT);
 
 IOTimeTypeStampS(IOINTES_INTFLTR);
     IOTimeStampConstant(IODBG_INTES(IOINTES_FILTER),
-                        (unsigned int) filterAction, (unsigned int) owner);
+                        (uintptr_t) filterAction, (uintptr_t) owner);
     filterRes = (*filterAction)(owner, this);
 IOTimeTypeStampE(IOINTES_INTFLTR);
 
@@ -180,7 +180,7 @@ IOTimeTypeStampS(IOINTES_INTCTXT);
 
 IOTimeTypeStampS(IOINTES_INTFLTR);
     IOTimeStampConstant(IODBG_INTES(IOINTES_FILTER),
-                        (unsigned int) filterAction, (unsigned int) owner);
+                        (uintptr_t) filterAction, (uintptr_t) owner);
     filterRes = (*filterAction)(owner, this);
 IOTimeTypeStampE(IOINTES_INTFLTR);
